Move collection-specific calls to collection manager
